feat(web-studio): add scheduled resource sync
