+++ web-studio/src/i18n/locales/en/resources.ts
@@ -1,0 +1,401 @@
+const resources = {
+  operations: {
+    page: {
+      placeholder: 'Operations dashboard is under construction.',
+    },
+  },
+  requestLogs: {
+    accessRequired: {
+      description:
+        'Current connection has no admin/root role. Configure an API key with Console Usage/Audit access in Connection & Identity.',
+      title: 'Admin access required',
+    },
+    clear: 'Clear',
+    description:
+      'Inspect server-side audited API requests for the current identity, including status, latency, and request identifiers.',
+    disabled: {
+      description:
+        'Usage/Audit is not initialized, so server-side request logs are unavailable.',
+      title: 'Audit logs unavailable',
+    },
+    empty: {
+      description: 'Start your first audited API call!',
+      filteredDescription:
+        'Adjust the query or status filter to broaden the visible log entries.',
+      filteredTitle: 'No matching requests',
+      title: 'No logs yet',
+      upload: 'Upload File',
+    },
+    error: {
+      description: 'Failed to load audited request logs from the server.',
+      title: 'Request failed',
+    },
+    eyebrow: 'Playground telemetry',
+    filters: {
+      all: 'All logs',
+      apiTypePlaceholder: 'API type',
+      error: 'Error logs',
+      requestIdPlaceholder: 'Exact Request ID',
+      statusCodePlaceholder: 'Status code',
+    },
+    loading: 'Loading request logs...',
+    metrics: {
+      successRate: 'Success rate',
+      total: 'Total calls',
+    },
+    pagination: {
+      next: 'Next',
+      pageSize: 'Rows per page',
+      pageSizeValue: '{{count}} / page',
+      previous: 'Previous',
+      summary: '{{total}} total, page {{page}} / {{pageCount}}',
+    },
+    query: 'Query',
+    refresh: 'Refresh',
+    reset: 'Reset',
+    searchPlaceholder: 'Filter method, path, or status',
+    scope: {
+      currentIdentity: 'Current scope: Current API key identity',
+      currentIdentityWithName:
+        'Current scope: Current API key identity ({{identity}})',
+    },
+    status: {
+      error: 'ERR',
+      pending: 'PENDING',
+      success: 'OK',
+    },
+    table: {
+      accountId: 'Account ID',
+      apiType: 'API Type',
+      duration: 'Duration',
+      method: 'Method',
+      path: 'Path',
+      requestId: 'Request ID',
+      status: 'Status',
+      time: 'Time',
+      title: 'Captured requests',
+      userId: 'User ID',
+    },
+    title: 'Request Logs',
+  },
+  addResource: {
+    title: 'Add Resource',
+    description:
+      'Upload a local file to the server. File type is auto-detected via magic bytes.',
+    dropzone: {
+      title: 'Drag & drop a file here, or click to select',
+      hint: 'Up to 10 files at a time.',
+      supportedFormats:
+        'Supports PDF, Word, PPTX, Excel, Markdown, code files, images, and more',
+    },
+    fileInfo: {
+      name: 'File',
+      size: 'Size',
+      type: 'Type',
+      unknown: 'Unknown type',
+      remove: 'Remove',
+    },
+    targetUri: 'Target URI',
+    'targetUri.placeholder': 'viking://resources/',
+    'targetUri.hint':
+      'Choose where to store this resource. Defaults to viking://resources/.',
+    'targetUri.browse': 'Browse',
+    advancedOptions: 'Advanced Options',
+    upload: 'Upload File',
+    'upload.processing': 'File uploaded, processing...',
+    uploading: 'Uploading\u2026',
+    result: {
+      success: 'Upload complete!',
+      skippedFiles: '{{count}} file(s) skipped (unsupported format)',
+    },
+    cancelUpload: 'Cancel',
+    startProcessing: 'Start Processing',
+    success: 'Resource added successfully',
+    fileBlocked: '"{{name}}" is not a supported file type.',
+    fileTooLarge: '"{{name}}" exceeds the {{size}} file size limit.',
+    tooManyFiles: 'Only the first {{count}} files were kept.',
+    error: 'Request Failed',
+    dirPicker: {
+      title: 'Select Directory',
+      select: 'Select',
+      cancel: 'Cancel',
+      empty: 'Empty directory',
+      error: 'Failed to load directory',
+      selected: 'Selected:',
+    },
+    mode: {
+      upload: 'Upload File',
+      remote: 'Remote URL',
+    },
+    remoteUrl: 'Remote URL',
+    'remoteUrl.placeholder': 'https://github.com/org/repo',
+    'remoteUrl.hint':
+      'HTTP(S) URL, Git repository, or other remote resource address.',
+    watch: {
+      enabled: 'Scheduled Sync',
+      hint: 'Periodically check this remote resource and import its latest content.',
+      interval: 'Sync interval (minutes)',
+      intervalHint: 'For example, 60 for hourly or 1440 for daily.',
+    },
+    strict: 'Strict Mode',
+    'strict.hint':
+      'When enabled, the server will reject files with unsupported or unrecognized types instead of skipping them silently.',
+    directlyUploadMedia: 'Directly Upload Media',
+    'directlyUploadMedia.hint':
+      'When enabled, media files (images, audio, video) are stored as-is. When disabled, media files are processed through AI vision/audio pipeline for content extraction first.',
+    createParent: 'Auto-create Parent Folder',
+    'createParent.hint':
+      'When enabled, automatically creates the parent directory if it does not exist.',
+    reason: 'Reason',
+    'reason.placeholder': 'Why are you adding this resource?',
+    instruction: 'Instruction',
+    'instruction.placeholder':
+      'Special processing instructions for this resource.',
+    directoryScan: {
+      title: 'Directory Scan Options',
+      ignoreDirs: 'Ignore Directories',
+      'ignoreDirs.placeholder': 'node_modules, .git, __pycache__',
+      include: 'Include Pattern',
+      'include.placeholder': '*.py, *.md',
+      exclude: 'Exclude Pattern',
+      'exclude.placeholder': '*.log, *.tmp',
+    },
+  },
+  resources: {
+    processingTasks: {
+      title: 'File Processing Tasks',
+      empty: 'No processing tasks',
+      toggleError: 'Toggle error details',
+      columns: {
+        fileName: 'File Name',
+        status: 'Status',
+        size: 'Size',
+      },
+      status: {
+        cancelled: 'Cancelled',
+        processing: 'Processing',
+        success: 'Processed',
+        failed: 'Processing failed',
+      },
+    },
+    searchPalette: {
+      ariaLabel: 'Search',
+      openContainingDirectory: 'Open containing directory',
+      placeholder: 'Search',
+      scope: {
+        global: 'Search scope: Global',
+        current: 'Search scope: {{name}}',
+        resetToGlobal: 'Click to reset to global search',
+      },
+      scopeState: {
+        validatingTitle: 'Validating search scope',
+        validatingPrefix: 'Checking whether',
+        validatingSuffix: 'exists',
+        switchTitle: 'Switch search scope',
+        switchPrefix: 'Press',
+        switchMiddle: 'to switch to',
+        invalidTitle: 'Search scope not found',
+        invalidPrefix: 'Path',
+        invalidSuffix: 'is inaccessible and cannot be switched to',
+      },
+      empty: {
+        title: 'Search files and directories',
+      },
+      browseDirHint: {
+        before: 'Enter',
+        after: 'to browse directories',
+      },
+      globalScopeHint: {
+        before: 'Enter',
+        after: 'to switch search scope to global',
+      },
+      error: 'Search failed',
+      emptyResults: {
+        title: 'No matching files or directories found',
+        subtitle: 'Try another keyword?',
+      },
+      footer: {
+        dirMode: {
+          select: 'Select',
+          level: 'Level',
+          confirm: 'Confirm',
+          cancel: 'Cancel',
+        },
+        resultMode: {
+          navigate: 'Navigate',
+          open: 'Open',
+          close: 'Close',
+          count: '{{count}} results',
+        },
+      },
+    },
+    dirBrowser: {
+      back: 'Back',
+      loading: 'Loading directory',
+      filesSection: 'Files',
+      error: 'Failed to load directory',
+      empty: {
+        title: 'Empty directory',
+        subtitle:
+          'There are currently no subdirectories to expand at this level',
+      },
+    },
+    filePreview: {
+      cancel: 'Cancel',
+      edit: 'Edit',
+      emptyFile: '(empty file)',
+      emptyPrompt: 'Select a file to preview it here',
+      formattingJson: 'Formatting JSON...',
+      imageFailed: 'Image failed to load.',
+      imageLoading: 'Loading image...',
+      largeFileSkipped: 'This file is large and was not loaded automatically.',
+      jsonFormatFailed: 'JSON formatting failed. Showing the raw content.',
+      loadFile: 'Load anyway',
+      loadingContent: 'Reading content...',
+      loadingEditor: 'Loading editor...',
+      markdownPreview: 'Preview',
+      markdownSource: 'Source',
+      noDirectoryContext: 'No abstract or overview available for this folder.',
+      save: 'Save',
+      selectDirectoryContext: 'Select a chip to show folder context.',
+      unsupportedBinary: 'Binary files do not support text preview.',
+      jsonl: {
+        collapse: 'Collapse',
+        dialogMode: 'Dialog',
+        emptyJsonl: 'Empty JSONL.',
+        emptyMessage: 'Empty message',
+        expand: 'Expand',
+        noArguments: 'No arguments',
+        rawMode: 'JSONL',
+        recordCount: '{{count}} record',
+        recordCount_other: '{{count}} records',
+        toolcall: 'toolcall',
+      },
+    },
+  },
+  retrieval: {
+    title: 'Retrieval',
+    searchPlaceholder: 'Search context',
+    placeholders: {
+      find: 'Enter a semantic query',
+      search: 'Enter a question to interpret with session context',
+      grep: 'Enter a regular expression or exact text',
+      glob: 'Enter a file pattern, for example **/*.md',
+    },
+    send: 'Search',
+    controls: {
+      function: 'Retrieval Function',
+      modes: {
+        find: 'find',
+        search: 'search',
+        grep: 'grep',
+        glob: 'glob',
+      },
+      ignoreCase: 'Ignore case',
+      resultCount: 'Results',
+      path: 'Path',
+      pathPlaceholder: '/',
+      scope: 'Scope',
+      customScope: 'Custom scope',
+      customScopePlaceholder: 'resources/project or viking://...',
+      effectiveScope: 'Scope',
+      allContexts: 'All contexts',
+      scopes: {
+        all: {
+          label: 'All contexts',
+        },
+        resources: {
+          label: 'Resources',
+        },
+        custom: {
+          label: 'Custom URI',
+        },
+      },
+      sessionId: 'Session ID',
+      sessionPlaceholder: 'session_id (optional)',
+    },
+    advanced: {
+      title: 'More filters',
+      contextTypes: 'Context types',
+      levels: 'Content levels',
+      scoreThreshold: 'Minimum relevance',
+      tags: 'Search tags',
+      tagsPlaceholder: 'For example env=prod, team=search',
+      tagsInvalid: 'Tags must use key=value format, separated by commas',
+      timeRange: 'Time range',
+      since: 'Start, e.g. 7d',
+      until: 'End, e.g. 2026-07-27',
+      anyTime: 'Any time',
+      quickRange: 'Quick range',
+      last24Hours: 'Last 24 hours',
+      last7Days: 'Last 7 days',
+      last30Days: 'Last 30 days',
+      customRange: 'Custom dates',
+      startDate: 'Start date',
+      endDate: 'End date',
+      clearTimeRange: 'Clear time range',
+      timeField: 'Time field',
+      updatedAt: 'Updated time',
+      createdAt: 'Created time',
+      provenance: 'Include retrieval provenance',
+      reset: 'Reset filters',
+    },
+    detail: {
+      score: 'Score',
+      level: 'Level',
+      openPlayground: 'Open in Playground',
+      uri: 'URI',
+      summary: 'Summary',
+      noSummary: 'No summary available',
+      matchReason: 'Match reason',
+    },
+    results: {
+      title: 'Search Results',
+      topN: 'Search Results (Top{{count}})',
+      line: 'Line {{line}}',
+      lineLabel: 'Line',
+      score: 'Score',
+      uri: 'URI',
+      description: 'Description',
+      provenance: 'Retrieval provenance ({{count}})',
+    },
+    types: {
+      resource: 'Resources',
+      memory: 'Memories',
+      skill: 'Skills',
+    },
+    queryPlan: {
+      title: '{{count}} planned queries',
+      more: '+{{count}} more',
+    },
+    loading: {
+      vector: 'Searching vector indexes...',
+      scan: 'Scanning context hierarchy...',
+      match: 'Matching semantic context...',
+      rerank: 'Reranking results...',
+    },
+    empty: {
+      checking: 'Checking retrievable context...',
+      readyTitle: 'Retrievable context is available',
+      readyDescription: 'Enter a keyword and press Enter to search',
+      title: 'No retrievable context yet',
+      description: 'Upload your first resource to get started.',
+      upload: 'Upload File',
+    },
+    error: {
+      title: 'Search failed',
+      fallback: 'The request could not be completed. Please try again.',
+      network:
+        'The OpenViking service could not be reached. Check the connection URL, service status, or CORS configuration.',
+      code: 'Error code',
+      status: 'HTTP status',
+      requestId: 'Request ID',
+    },
+    noResults: {
+      title: 'No matching content found',
+      subtitle: 'Try another keyword or adjust the path scope',
+    },
+  },
+} as const
+
+export default resources
