+++ web-studio/src/i18n/locales/zh-CN/resources.ts
@@ -1,0 +1,391 @@
+const resources = {
+  operations: {
+    page: {
+      placeholder: '运维面板能力尚未接入。',
+    },
+  },
+  requestLogs: {
+    accessRequired: {
+      description:
+        '当前连接未获取到 admin/root 权限，无法显示请求日志。请在连接与身份中配置具备 Console Usage/Audit 权限的 API Key。',
+      title: '需要 admin 权限',
+    },
+    clear: '清空',
+    description:
+      '查看当前身份下服务端审计到的 API 请求，包括状态、耗时和请求标识。',
+    disabled: {
+      description: 'Usage/Audit 未初始化，暂无服务端请求日志。',
+      title: '审计日志不可用',
+    },
+    empty: {
+      description: '先开始您的第一次可审计调用吧！',
+      filteredDescription: '调整搜索内容或状态筛选，扩大可见日志范围。',
+      filteredTitle: '没有匹配的请求',
+      title: '当前无日志信息',
+      upload: '上传文件',
+    },
+    error: {
+      description: '无法从服务端加载审计请求日志。',
+      title: '请求失败',
+    },
+    eyebrow: 'Playground 遥测',
+    filters: {
+      all: '所有日志',
+      apiTypePlaceholder: 'API 类型',
+      error: '错误日志',
+      requestIdPlaceholder: '精确 Request ID',
+      statusCodePlaceholder: '状态码',
+    },
+    loading: '正在加载请求日志...',
+    metrics: {
+      successRate: '成功率',
+      total: '总调用次数',
+    },
+    pagination: {
+      next: '下一页',
+      pageSize: '每页条数',
+      pageSizeValue: '每页 {{count}} 条',
+      previous: '上一页',
+      summary: '共 {{total}} 条，第 {{page}} / {{pageCount}} 页',
+    },
+    query: '查询',
+    refresh: '刷新',
+    reset: '重置',
+    searchPlaceholder: '筛选方法、路径或状态码',
+    scope: {
+      currentIdentity: '当前范围：当前 API key 身份',
+      currentIdentityWithName: '当前范围：当前 API key 身份（{{identity}}）',
+    },
+    status: {
+      error: 'ERR',
+      pending: 'PENDING',
+      success: 'OK',
+    },
+    table: {
+      accountId: 'Account ID',
+      apiType: 'API 类型',
+      duration: '耗时',
+      method: '方法',
+      path: '路径',
+      requestId: 'Request ID',
+      status: '状态',
+      time: '时间',
+      title: '捕获的请求',
+      userId: 'User ID',
+    },
+    title: '请求日志',
+  },
+  addResource: {
+    title: '添加资源',
+    description: '上传本地文件到服务器，文件类型通过 magic bytes 自动检测。',
+    dropzone: {
+      title: '拖拽文件到此处，或点击选择文件',
+      hint: '每次最多上传 10 个文件。',
+      supportedFormats:
+        '支持 PDF、Word、PPTX、Excel、Markdown、代码文件、图片等',
+    },
+    fileInfo: {
+      name: '文件',
+      size: '大小',
+      type: '类型',
+      unknown: '未知类型',
+      remove: '移除',
+    },
+    targetUri: '目标 URI',
+    'targetUri.placeholder': 'viking://resources/',
+    'targetUri.hint': '选择资源的存储位置，默认为 viking://resources/。',
+    'targetUri.browse': '浏览',
+    advancedOptions: '高级选项',
+    upload: '上传文件',
+    'upload.processing': '文件已上传，正在处理中...',
+    uploading: '上传中…',
+    result: {
+      success: '上传完成！',
+      skippedFiles: '{{count}} 个文件被跳过（不支持的格式）',
+    },
+    cancelUpload: '取消',
+    startProcessing: '开始处理',
+    success: '资源添加成功',
+    fileBlocked: '"{{name}}" 不是支持的文件类型。',
+    fileTooLarge: '"{{name}}" 超过 {{size}} 文件大小限制。',
+    tooManyFiles: '仅保留前 {{count}} 个文件，其余已忽略。',
+    error: '请求失败',
+    dirPicker: {
+      title: '选择目录',
+      select: '选择',
+      cancel: '取消',
+      empty: '空目录',
+      error: '加载目录失败',
+      selected: '已选择：',
+    },
+    mode: {
+      upload: '上传文件',
+      remote: '远程资源',
+    },
+    remoteUrl: '远程资源地址',
+    'remoteUrl.placeholder': 'https://github.com/org/repo',
+    'remoteUrl.hint': 'HTTP(S) 链接、Git 仓库地址或其他远程资源地址。',
+    watch: {
+      enabled: '定时同步',
+      hint: '定期检查该远程资源，并导入最新内容。',
+      interval: '同步周期（分钟）',
+      intervalHint: '例如每小时填写 60，每天填写 1440。',
+    },
+    strict: '严格模式',
+    'strict.hint':
+      '开启时，服务器会拒绝不支持或无法识别类型的文件，而非静默跳过。',
+    directlyUploadMedia: '直接上传媒体文件',
+    'directlyUploadMedia.hint':
+      '开启时，媒体文件（图片、音频、视频）原样存储。关闭后，媒体文件会先通过 AI 视觉/音频管道提取内容再存储。',
+    createParent: '自动创建父文件夹',
+    'createParent.hint': '开启时，若目标父目录不存在则自动创建。',
+    reason: '添加原因',
+    'reason.placeholder': '为什么要添加这个资源？',
+    instruction: '处理指令',
+    'instruction.placeholder': '针对该资源的特殊处理指令。',
+    directoryScan: {
+      title: '目录扫描选项',
+      ignoreDirs: '忽略目录',
+      'ignoreDirs.placeholder': 'node_modules, .git, __pycache__',
+      include: '包含模式',
+      'include.placeholder': '*.py, *.md',
+      exclude: '排除模式',
+      'exclude.placeholder': '*.log, *.tmp',
+    },
+  },
+  resources: {
+    processingTasks: {
+      title: '文件处理任务',
+      empty: '暂无处理任务',
+      toggleError: '展开或收起错误详情',
+      columns: {
+        fileName: '文件名',
+        status: '状态',
+        size: '大小',
+      },
+      status: {
+        cancelled: '已取消',
+        processing: '处理中',
+        success: '处理成功',
+        failed: '处理失败',
+      },
+    },
+    searchPalette: {
+      ariaLabel: '搜索',
+      openContainingDirectory: '打开所在目录',
+      placeholder: '搜索',
+      scope: {
+        global: '搜索范围: 全局',
+        current: '搜索范围: {{name}}',
+        resetToGlobal: '点击重置为全局搜索',
+      },
+      scopeState: {
+        validatingTitle: '正在校验搜索范围',
+        validatingPrefix: '正在检查',
+        validatingSuffix: '是否存在',
+        switchTitle: '切换搜索范围',
+        switchPrefix: '按',
+        switchMiddle: '切换到',
+        invalidTitle: '搜索范围不存在',
+        invalidPrefix: '路径',
+        invalidSuffix: '无法访问，不能切换',
+      },
+      empty: {
+        title: '搜索文件和目录',
+      },
+      browseDirHint: {
+        before: '输入',
+        after: '浏览目录结构',
+      },
+      globalScopeHint: {
+        before: '输入',
+        after: '切换搜索范围到全局',
+      },
+      error: '搜索出错',
+      emptyResults: {
+        title: '没有找到匹配的文件或目录',
+        subtitle: '试试换个关键词？',
+      },
+      footer: {
+        dirMode: {
+          select: '选择',
+          level: '层级',
+          confirm: '确定',
+          cancel: '取消',
+        },
+        resultMode: {
+          navigate: '导航',
+          open: '打开',
+          close: '关闭',
+          count: '{{count}} 个结果',
+        },
+      },
+    },
+    dirBrowser: {
+      back: '返回上一级',
+      loading: '正在加载目录',
+      filesSection: '文件',
+      error: '加载目录失败',
+      empty: {
+        title: '空目录',
+        subtitle: '这一层目前没有可继续展开的子目录',
+      },
+    },
+    filePreview: {
+      cancel: '取消',
+      edit: '编辑',
+      emptyFile: '(空文件)',
+      emptyPrompt: '选择文件后在这里预览',
+      formattingJson: '正在格式化 JSON...',
+      imageFailed: '图片加载失败。',
+      imageLoading: '正在加载图片...',
+      largeFileSkipped: '文件较大，默认不自动加载。',
+      jsonFormatFailed: 'JSON 格式化失败，已显示原始内容。',
+      loadFile: '仍然加载',
+      loadingContent: '正在读取内容...',
+      loadingEditor: '加载编辑器...',
+      markdownPreview: '预览',
+      markdownSource: '源码',
+      noDirectoryContext: '这个目录暂无摘要或概览。',
+      save: '保存',
+      selectDirectoryContext: '选择一个标签查看目录上下文。',
+      unsupportedBinary: '二进制文件不支持文本预览。',
+      jsonl: {
+        collapse: '收起',
+        dialogMode: '对话',
+        emptyJsonl: '空 JSONL。',
+        emptyMessage: '空消息',
+        expand: '展开',
+        noArguments: '无参数',
+        rawMode: 'JSONL',
+        recordCount: '{{count}} 条记录',
+        toolcall: '工具调用',
+      },
+    },
+  },
+  retrieval: {
+    title: '检索',
+    searchPlaceholder: '输入检索内容',
+    placeholders: {
+      find: '输入语义检索内容',
+      search: '输入需要结合会话理解的问题',
+      grep: '输入正则表达式或精确文本',
+      glob: '输入文件模式，例如 **/*.md',
+    },
+    send: '检索',
+    controls: {
+      function: '检索函数',
+      modes: {
+        find: 'find',
+        search: 'search',
+        grep: 'grep',
+        glob: 'glob',
+      },
+      ignoreCase: '忽略大小写',
+      resultCount: '返回数量',
+      path: '路径',
+      pathPlaceholder: '/',
+      scope: '检索范围',
+      customScope: '自定义范围',
+      customScopePlaceholder: 'resources/project 或 viking://...',
+      effectiveScope: '范围',
+      allContexts: '全部上下文',
+      scopes: {
+        all: {
+          label: '全部上下文',
+        },
+        resources: {
+          label: '资源库',
+        },
+        custom: {
+          label: '自定义 URI',
+        },
+      },
+      sessionId: 'Session ID',
+      sessionPlaceholder: 'session_id（可选）',
+    },
+    advanced: {
+      title: '更多筛选',
+      contextTypes: '上下文类型',
+      levels: '内容层级',
+      scoreThreshold: '最低相关度',
+      tags: '检索标签',
+      tagsPlaceholder: '例如 env=prod, team=search',
+      tagsInvalid: '标签必须使用 key=value 格式，多个标签用逗号分隔',
+      timeRange: '时间范围',
+      since: '开始时间，如 7d',
+      until: '结束时间，如 2026-07-27',
+      anyTime: '不限时间',
+      quickRange: '快捷范围',
+      last24Hours: '近 24 小时',
+      last7Days: '近 7 天',
+      last30Days: '近 30 天',
+      customRange: '自定义日期',
+      startDate: '开始日期',
+      endDate: '结束日期',
+      clearTimeRange: '清除时间范围',
+      timeField: '时间字段',
+      updatedAt: '更新时间',
+      createdAt: '创建时间',
+      provenance: '返回检索过程',
+      reset: '重置筛选',
+    },
+    detail: {
+      score: '相关度',
+      level: '层级',
+      openPlayground: '在 Playground 打开',
+      uri: 'URI',
+      summary: '摘要',
+      noSummary: '暂无可展示的摘要',
+      matchReason: '匹配原因',
+    },
+    results: {
+      title: '检索结果',
+      topN: '检索结果（Top{{count}}）',
+      line: '第 {{line}} 行',
+      lineLabel: '行号',
+      score: '相关度',
+      uri: 'URI',
+      description: '描述',
+      provenance: '检索过程（{{count}}）',
+    },
+    types: {
+      resource: 'Resources',
+      memory: 'Memories',
+      skill: 'Skills',
+    },
+    queryPlan: {
+      title: '查询计划 {{count}} 条',
+      more: '+{{count}} 条',
+    },
+    loading: {
+      vector: '正在检索向量索引...',
+      scan: '扫描知识库层级结构...',
+      match: '匹配语义相关内容...',
+      rerank: '对结果重排序...',
+    },
+    empty: {
+      checking: '正在检查可检索上下文...',
+      readyTitle: '已有可检索上下文',
+      readyDescription: '输入关键词后按 Enter 开始检索',
+      title: '当前还没有可检索的上下文',
+      description: '先上传您的第一份资源吧～',
+      upload: '上传文件',
+    },
+    error: {
+      title: '检索失败',
+      fallback: '请求未能完成，请稍后重试。',
+      network:
+        '请求未能到达 OpenViking 服务，请检查连接地址、服务状态或跨域配置。',
+      code: '错误码',
+      status: 'HTTP 状态',
+      requestId: 'Request ID',
+    },
+    noResults: {
+      title: '没有找到匹配的内容',
+      subtitle: '试试换个关键词或调整路径范围',
+    },
+  },
+} as const
+
+export default resources
