+++ web-studio/src/lib/watch-interval.ts
@@ -1,0 +1,4 @@
+export function parsePositiveMinutes(value: number | string): number | null {
+  const minutes = Number(value)
+  return Number.isFinite(minutes) && minutes > 0 ? minutes : null
+}
